Move XML related code to a separate folder
