+++ src/cpp/shared/xml/campaign_xml.h
