+++ src/cpp/shared/xml/ironfist_xml.h
