+++ src/vcxproj/ironfist/ironfist.vcxproj.filters
@@ -91,6 +91,12 @@
     <Filter Include="Header Files\campaign">
       <UniqueIdentifier>{dfdbcac3-d0b6-4cb6-8459-3d5559be70f0}</UniqueIdentifier>
     </Filter>
+    <Filter Include="Source Files\xml">
+      <UniqueIdentifier>{c3d5fe7e-1828-4db9-88f0-67391d167a2f}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Header Files\xml">
+      <UniqueIdentifier>{2bbf2978-e172-4be9-82fc-3f2f8f983330}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\cpp\shared\init.cpp">
@@ -321,11 +327,11 @@
     <ClCompile Include="..\..\cpp\shared\adventure\hero_globals.cpp">
       <Filter>Source Files\adventure</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\cpp\shared\game\ironfist_xml.cpp">
-      <Filter>Source Files\game</Filter>
+    <ClCompile Include="..\..\cpp\shared\xml\campaign_xml.cpp">
+      <Filter>Source Files\xml</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\cpp\shared\campaign\campaign_xml.cpp">
-      <Filter>Source Files\campaign</Filter>
+    <ClCompile Include="..\..\cpp\shared\xml\ironfist_xml.cpp">
+      <Filter>Source Files\xml</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -551,14 +557,14 @@
     <ClInclude Include="..\..\cpp\shared\smacker.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\cpp\shared\game\ironfist_xml.h">
-      <Filter>Header Files\game</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\cpp\shared\gui\msg.h">
       <Filter>Header Files\gui</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\cpp\shared\campaign\campaign_xml.h">
-      <Filter>Header Files\campaign</Filter>
+    <ClInclude Include="..\..\cpp\shared\xml\campaign_xml.h">
+      <Filter>Header Files\xml</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\cpp\shared\xml\ironfist_xml.h">
+      <Filter>Header Files\xml</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
